chore: Update npm dependencies
This commit updates the npm dependencies in the `package.json` and `package-lock.json` files. It includes updating the following dependencies:
- `next` to version 14.2.3
- `prettier` to version 3.2.5
- `eslint` to version 8
- `eslint-config-next` to version 14.2.3
- `tailwindcss` to version 3.4.1
- `typescript` to version 5

These updates were made to ensure that the project is using the latest stable versions of the dependencies and to take advantage of any bug fixes or new features introduced in these versions.